T32 issue
The T32 trigger was checking only the first slot signal (2 byte). It has to check all the 4 bytes of the signal otherwise it is not able to undestand correctly the new signal (i.e. if the signal is different from the previous for the temperature only that is on the last 2 bytes, the trigger will not be activated)